Add GitHub Actions workflow to publish files to production (#1719)
This will add a workflow action that will take a comma separated list of
files as input (or "ALL" by default if all files are desired) and create
a pull-request for updating the content of the files and all of the
translations thereof. After reviewing the PR it can be merged to publish
all of the file changes to the production website.