Update internal links from list at top of ToS and Privacy policy to be correct
This means the internal-page links are different from when they were hand-cut
HTML, but a user following an old link will then come to the top of the page
and will get the list of in-page links, so should be able to easily find
the relevant content. The odds of mid-page links in the wild to these two
documents are hopefully pretty small.