Create Formulaic Report a Violation
Please create a page for Formulaic to share how users can report a violation.  There is not a URL for this page yet, but we will have one created once the document is put together.  Note that in line 33, "Terms of Service" and "Content Policies" will eventually need to include links to the pages with the Formulaic TOS and Content Policies.